chore: Rename QueryBuilder to VisualQueryBuilder
