+++ src/components/query/visual-query-builder/PropertyQueryEditor.tsx
@@ -16,8 +16,8 @@ import {
 } from 'types';
 import { LinkButton, Select, Input, Icon } from '@grafana/ui';
 import { SitewiseQueryEditorProps } from './types';
-import { AssetBrowser } from '../browser/AssetBrowser';
-import { AggregatePicker, aggReg } from '../AggregatePicker';
+import { AssetBrowser } from '../../browser/AssetBrowser';
+import { AggregatePicker, aggReg } from '../../AggregatePicker';
 import { getAssetProperty, getDefaultAggregate } from 'queryInfo';
 import { QualityAndOrderRow } from './QualityAndOrderRow';
 import { EditorField, EditorFieldGroup, EditorRow } from '@grafana/experimental';
